Test createArtwork controller method
