+++ src/artwork/controller/types.ts
@@ -1,4 +1,5 @@
 import { type NextFunction, type Request, type Response } from "express";
+import { type ArtworkData } from "../repository/types";
 
 export interface ArtworksControllerStructure {
   getArtworks(req: Request, res: Response, next: NextFunction): void;
@@ -6,3 +7,9 @@ export interface ArtworksControllerStructure {
 }
 
 export type ResponseWithStatusJson = Pick<Response, "status" | "json">;
+
+export type RequestWithArtworkData = Request<
+  Record<string, unknown>,
+  Record<string, unknown>,
+  ArtworkData
+>;
